gen4 commands refactoring
